% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arulePredict.R
\name{arulePredict}
\alias{arulePredict}
\title{Creates recommendations based on input association rules}
\usage{
arulePredict(rules, newdata, customerkey, productkey)
}
\arguments{
\item{rules}{Association Rule Mining Training Vector Object}

\item{newdata}{New input data.frame with customers and items}

\item{customerkey}{The name of the Column containing customer ID's}

\item{productkey}{The name of the Column containing item/product ID's}
}
\description{
This function returns recommendations based on input association rules and a new unseen dataset
}
\examples{
rule_predict(rules, newdata, customerkey, productkey)
}
