% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aruleRec.R
\name{aruleRec}
\alias{aruleRec}
\title{Easy and fast recommendations with association rule learning in R}
\usage{
aruleRec(data, productkey, customerkey, support = 0.01,
  confidence = 0.7, minlen = 2,
  maxlen = as.integer(length(unique(data$productkey)) + 1))
}
\arguments{
\item{data}{A dataframe containing dataframe as input in tidy format with one customer-item pair per row. This data.frame must contain two columns: a customer ID and a product ID.}

\item{productkey}{The column containing item ID's}

\item{customerkey}{The column containing customer ID's}

\item{support}{Minimum support of the mined association rules}

\item{confidence}{Minimum confidence of the mined association rules}

\item{minlen}{Minimum length of the mined association rules}

\item{maxlen}{Maximum length of the mined association rules}
}
\description{
The aruleRec function generates recommendations from an input dataframe.
}
\examples{
aruleRec(data = dat, productkey = ProductKey, customerkey = CustomerKey, minlen = 2, maxlen = 20, support = 0.01, confidence = 0.7)
}
