% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aruleTrain.R
\name{aruleTrain}
\alias{aruleTrain}
\title{Mines association rules from an input data.frame}
\usage{
aruleTrain(data, productkey, customerkey, support = 0.01,
  confidence = 0.7, minlen = 2,
  maxlen = as.integer(length(unique(data$productkey)) + 1))
}
\arguments{
\item{data}{A dataframe containing dataframe as input in tidy format with one customer-item pair per row. This data.frame must contain two columns: a customer ID and a product ID.}

\item{productkey}{The column containing item ID's}

\item{customerkey}{The column containing customer ID's}

\item{support}{Minimum support of the mined association rules}

\item{confidence}{Minimum confidence of the mined association rules}

\item{minlen}{Minimum length of the mined association rules}

\item{maxlen}{Maximum length of the mined association rules}
}
\description{
This function mnes association rules from an input data.frame but does nothing more
}
\examples{
aruleTrain(data, productkey, customerkey, cupport, confidence, minlen, maxlen)
}
