% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_items_by_contact.R
\name{create_items_by_contact}
\alias{create_items_by_contact}
\title{Create Signups By Contact}
\usage{
create_items_by_contact(data, customerkey, productkey)
}
\arguments{
\item{data}{A dataframe containing dataframe as input in tidy format with one customer-item pair per row. This data.frame must contain two columns: a customer ID and a product ID.}

\item{customerkey}{The column containing customer ID's}

\item{productkey}{The column containing item ID's}
}
\description{
This creates a new dataframe with items by contacts
}
\examples{
create_items_by_contact(data, customerkey, productkey)
}
