% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rules.R
\name{create_rules}
\alias{create_rules}
\title{Mines the association rules for the subsequent recommendations}
\usage{
create_rules(transactions, support, confidence, minlen, maxlen)
}
\arguments{
\item{transactions}{A transaction object}

\item{support}{Minimum support of the mined association rules}

\item{confidence}{Minimum confidence of the mined association rules}

\item{minlen}{Minimum length of the mined association rules}

\item{maxlen}{Maximum length of the mined association rules}
}
\description{
This function creates transactions object required for association rule mining
}
\examples{
create_rules(transactions, support, confidence, minlen, maxlen)
}
