% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transactions.R
\name{create_transactions}
\alias{create_transactions}
\title{Create transactions object required for association rule mining}
\usage{
create_transactions(items_by_contact, productkey, customerkey)
}
\arguments{
\item{items_by_contact}{A dataframe containing items by contact}

\item{productkey}{The column containing item ID's}

\item{customerkey}{The column containing customer ID's}
}
\description{
This function creates transactions object required for association rule mining
}
\examples{
ccreate_transactions(items_by_contact, productkey, customerkey)
}
