% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_function.R
\name{export_function}
\alias{export_function}
\title{Exports obejcts to globalenv}
\usage{
export_function(...)
}
\arguments{
\item{...}{Stuff that needs to be exported}
}
\description{
Exports obejcts to globalenv
}
\examples{
export_function(data)
}
