% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_flow.R
\name{rule_flow}
\alias{rule_flow}
\title{The main rule flow function}
\usage{
rule_flow(data, customerkey, productkey, support, confidence, minlen,
  maxlen)
}
\arguments{
\item{data}{A dataframe containing dataframe as input in tidy format with one customer-item pair per row. This data.frame must contain two columns: a customer ID and a product ID.}

\item{customerkey}{The column containing customer ID's}

\item{productkey}{The column containing item ID's}

\item{support}{Minimum support of the mined association rules}

\item{confidence}{Minimum confidence of the mined association rules}

\item{minlen}{Minimum length of the mined association rules}

\item{maxlen}{Maximum length of the mined association rules}
}
\description{
This function mines the association rules and preprocesses data by calling other functions in the library.
}
\examples{
rule_flow(data, customerkey, productkey, support, confidence, minlen, maxlen)
}
